move data to user
