+++ src/app/api/participant/route.js
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server";
 import { db } from "../../../../firebase";
 import { doc, updateDoc } from "firebase/firestore";
 import { authenticate } from "@/utils/auth";
-import { AUTH } from "@/data/dynamic/admin/Participant";
+import { AUTH } from "@/data/dynamic/user/Participant";
 
 export async function POST(req) {
   const res = NextResponse;
